+++ src/main/java/com/programmers/application/controller/console/voucher/command/VoucherCommand.java
@@ -1,0 +1,8 @@
+package com.programmers.application.controller.console.voucher.command;
+
+public enum VoucherCommand {
+    EXIT,
+    CREATE,
+    LIST,
+    ;
+}
