--- src/main/java/com/programmers/application/controller/voucher/command/VoucherCommand.java
@@ -1,8 +1,0 @@
-package com.programmers.application.controller.voucher.command;
-
-public enum VoucherCommand {
-    EXIT,
-    CREATE,
-    LIST,
-    ;
-}
